third_party/WebKit/LayoutTests/css3/device-adapt/opera/constrain-011-expected.txt


FAIL CSS Test: @viewport constrained - zoom: 200% not affecting width assert_equals: expected 320 but got 160
